ChatLogMessageCellwを別ファイルに切り分け
